remove old gui
