+++ RayZath.sln
@@ -1,14 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30406.217
+# Visual Studio Version 17
+VisualStudioVersion = 17.2.32602.215
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "RayZath", "RayZath\RayZath.vcxproj", "{95A5C7CA-CBEF-4824-B977-E2821522510A}"
-EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Tester", "Tester\Tester.vcxproj", "{C31A1E26-FE8F-43B7-8E77-80D06F76133F}"
-	ProjectSection(ProjectDependencies) = postProject
-		{95A5C7CA-CBEF-4824-B977-E2821522510A} = {95A5C7CA-CBEF-4824-B977-E2821522510A}
-	EndProjectSection
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "UI", "UI\UI.vcxproj", "{E103CF3A-A86A-457F-8BEA-2F409609A2F5}"
 	ProjectSection(ProjectDependencies) = postProject
@@ -33,14 +28,6 @@ Global
 		{95A5C7CA-CBEF-4824-B977-E2821522510A}.Release|x64.ActiveCfg = Release|x64
 		{95A5C7CA-CBEF-4824-B977-E2821522510A}.Release|x64.Build.0 = Release|x64
 		{95A5C7CA-CBEF-4824-B977-E2821522510A}.Release|x86.ActiveCfg = Release|x64
-		{C31A1E26-FE8F-43B7-8E77-80D06F76133F}.Debug|x64.ActiveCfg = Debug|x64
-		{C31A1E26-FE8F-43B7-8E77-80D06F76133F}.Debug|x64.Build.0 = Debug|x64
-		{C31A1E26-FE8F-43B7-8E77-80D06F76133F}.Debug|x86.ActiveCfg = Debug|Win32
-		{C31A1E26-FE8F-43B7-8E77-80D06F76133F}.Debug|x86.Build.0 = Debug|Win32
-		{C31A1E26-FE8F-43B7-8E77-80D06F76133F}.Release|x64.ActiveCfg = Release|x64
-		{C31A1E26-FE8F-43B7-8E77-80D06F76133F}.Release|x64.Build.0 = Release|x64
-		{C31A1E26-FE8F-43B7-8E77-80D06F76133F}.Release|x86.ActiveCfg = Release|Win32
-		{C31A1E26-FE8F-43B7-8E77-80D06F76133F}.Release|x86.Build.0 = Release|Win32
 		{E103CF3A-A86A-457F-8BEA-2F409609A2F5}.Debug|x64.ActiveCfg = Debug|x64
 		{E103CF3A-A86A-457F-8BEA-2F409609A2F5}.Debug|x64.Build.0 = Debug|x64
 		{E103CF3A-A86A-457F-8BEA-2F409609A2F5}.Debug|x86.ActiveCfg = Debug|Win32
